[LeetCode]  637. Average of Levels in Binary Tree
쓸 데 없이 double로 선언한 변수 int로 다시 선언.
nt형은 double로 자동형변환되므로 굳이 쓸 데 없이 size를 double형으로 선언할 필요가 없습니다.

+++ LeetCode/BFS/637. Average of Levels in Binary Tree/Ingyu.java
@@ -1,9 +1,3 @@
-import javax.swing.tree.TreeNode;
-import java.util.ArrayList;
-import java.util.LinkedList;
-import java.util.List;
-import java.util.Queue;
-
 class Solution {
     public List<Double> averageOfLevels(TreeNode root) {
         List<Double> answerList = new ArrayList<>();
@@ -11,8 +5,8 @@ class Solution {
         q.offer(root);
 
         while(!q.isEmpty()) {
-            double size = q.size();
-            int temp = (int) size;
+            int size = q.size();
+            int temp = size;
             double sum = 0;
             while (temp-- > 0) {
                 TreeNode node = q.poll();
